animation-glib: Move vector.cpp/h out into animation-glib toplevel
We'll be using this in other subdirectories
