+++ animation-glib/wobbly/model.h
@@ -25,7 +25,7 @@
 #include <glib-object.h>
 
 #include <animation-glib/wobbly/anchor.h>
-#include <animation-glib/wobbly/vector.h>
+#include <animation-glib/vector.h>
 
 G_BEGIN_DECLS
